+++ CHANGELOG.md
@@ -7,6 +7,12 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 <!-- TODO: add @added tag to the files with specific version -->
 <!--       during new scriptlets or redirects releasing -->
 
+## [Unreleased]
+
+### Added
+
+- ability to use flags in RegExp [#303](https://github.com/AdguardTeam/Scriptlets/issues/303)
+
 ## <a name="v1.9.37"></a> [v1.9.37] - 2023-06-06
 
 ### Added